# [CAN-3237] Fixes how body & header keys look like on json logs