--- a/lib/formatters/JsonFormatter.js
+++ b/lib/formatters/JsonFormatter.js
@@ -1,3 +1,4 @@
+const { merge } = require('lodash');
 const BaseFormatter = require('./BaseFormatter');
 
 class JsonFormatter extends BaseFormatter {
@@ -60,11 +61,10 @@ class JsonFormatter extends BaseFormatter {
         id: obj.requestId
       }
     };
-
+    const meta = merge(metaBody, metaHeaders);
     return {
       ...defaultObject,
-      ...metaBody,
-      ...metaHeaders,
+      ...meta,
       ...remainingProperties
     };
   }