--- a/lib/formatters/SimpleFormatter.js
+++ b/lib/formatters/SimpleFormatter.js
@@ -1,5 +1,6 @@
 const BaseFormatter = require('./BaseFormatter');
 const chalk = require('chalk');
+const flat = require('flat');
 const serialize = require('../serializers/toKeyValue');
 
 class SimpleFormatter extends BaseFormatter {
@@ -19,6 +20,12 @@ class SimpleFormatter extends BaseFormatter {
       requestId,
       ...rest
     } = obj;
+    if (rest.metaHeaders) {
+      rest.metaHeaders = flat(rest.metaHeaders, { safe: true });
+    }
+    if (rest.metaBody) {
+      rest.metaBody = flat(rest.metaBody, { safe: true });
+    }
     if (options && options.traceHeaderName) {
       rest.metaHeaders && delete rest.metaHeaders['headers.' + options.traceHeaderName.toLowerCase()];
     }