--- a/lib/loggable.js
+++ b/lib/loggable.js
@@ -1,5 +1,4 @@
 const EventEmitter = require('events');
-const flat = require('flat');
 
 const EVENT = {
   INBOUND_REQUEST: 'inbound_request',