--- a/lib/transformers/transformers.js
+++ b/lib/transformers/transformers.js
@@ -1,5 +1,4 @@
 const logTags = require('../logTags');
-const flat = require('flat');
 const pick = require('lodash/pick');
 const pickBy = require('lodash/pickBy');
 const url = require('url');
@@ -91,7 +90,7 @@ const mapOutReq = (requestOptions, reqId, opts = {}) => {
         picked = pick(jsonObject, opts.bodyKeys);
       }
       if (Object.keys(picked).length) {
-        metaBody = flat({ body: picked }, { safe: true });
+        metaBody = { body: picked };
       }
     } catch (e) {}
   }
@@ -160,7 +159,11 @@ function extractRequestMeta(ctx, bodyKeys, bodyKeysRegex, headersRegex, prefix =
       picked = pick(ctx.request.body, bodyKeys);
     }
     if (Object.keys(picked).length) {
-      metaBody = flat({ [`${prefix + (prefix && '.')}body`]: picked }, { safe: true });
+      if (prefix) {
+        metaBody = { [prefix]: { body: picked } };
+      } else {
+        metaBody = { body: picked };
+      }
     }
   }