--- a/lib/utils.js
+++ b/lib/utils.js
@@ -1,5 +1,3 @@
-const flat = require('flat');
-
 function safeExec(fn, logger) {
   try {
     return fn();
@@ -22,7 +20,11 @@ function extractHeaders(headersRegex, headers, prefix = '') {
   });
 
   if (Object.keys(metaHeaders).length) {
-    metaHeaders = flat({ [`${prefix + (prefix && '.')}headers`]: metaHeaders });
+    if (prefix) {
+      metaHeaders = { [prefix]: { headers: metaHeaders } };
+    } else {
+      metaHeaders = { headers: metaHeaders };
+    }
   }
 
   return metaHeaders;