--- a/test/lib/formatters/ExtendedFormatterTest.js
+++ b/test/lib/formatters/ExtendedFormatterTest.js
@@ -15,7 +15,21 @@ describe('Test Extended Formatter', () => {
   it('formatOsbject should return base prefix', () => {
     const obj = { method: 'GET', requestId: 'testRequestId', path: 'test-path', log_tag: 'inbound_request' };
     const payload = this.extendedFormatter.formatObject(obj);
-
     payload.should.containEql('Test Prefix |');
+  });
+
+  it('formatObject should return object with flatten header & body keys', () => {
+    const obj = {
+      method: 'GET',
+      requestId: 'testRequestId',
+      path: 'test-path',
+      log_tag: 'inbound_request',
+      metaBody: { body: { fruit: 'banana', items: ['apple', 'orange'] } },
+      metaHeaders: { headers: { 'x-custom-header': 'test' } }
+    };
+    const payload = this.extendedFormatter.formatObject(obj);
+    payload.should.containEql('body.fruit="banana"');
+    payload.should.containEql('body.items=["apple", "orange"]');
+    payload.should.containEql('headers.x-custom-header="test"');
   });
 });