--- a/test/lib/formatters/SimpleFormatterTest.js
+++ b/test/lib/formatters/SimpleFormatterTest.js
@@ -18,4 +18,19 @@ describe('Test Simple Formatter', () => {
 
     payload.should.equal('Test Prefix');
   });
+
+  it('formatObject should return object with flatten header & body keys', () => {
+    const obj = {
+      method: 'GET',
+      requestId: 'testRequestId',
+      path: 'test-path',
+      log_tag: 'inbound_request',
+      metaBody: { body: { fruit: 'banana', items: ['apple', 'orange'] } },
+      metaHeaders: { headers: { 'x-custom-header': 'test' } }
+    };
+    const payload = this.simpleFormatter.formatObject(obj);
+    payload.should.containEql('body.fruit="banana"');
+    payload.should.containEql('body.items=["apple", "orange"]');
+    payload.should.containEql('headers.x-custom-header="test"');
+  });
 });